refactor: extract background sync runtime service
